Live announcements on bot startup added

+++ src/main/java/com/StreamerSpectrum/BeamTeamDiscordBot/singletons/JDAManager.java
@@ -162,7 +162,9 @@ public abstract class JDAManager {
 	}
 
 	public static void deleteMessage(GoLiveMessage message) {
-		deleteMessage(message.getMessageID(), message.getGoLiveChannelID());
+		if (GuildManager.getGuild(message.getGuildID()).isRemoveOfflineChannelAnnouncements()) {
+			deleteMessage(message.getMessageID(), message.getGoLiveChannelID());
+		}
 	}
 
 	public static void deleteMessage(String messageID, String goLiveChannelID) {
